🔨 Refactor: 서비스 레이어 추상화를 통한 의존성 역전(DIP) 적용
- 코드 분리 작업

Closes #14
